# src/main/resources/codegen/swagger.yml
openapi: '3.0.0'
info:
  title: 'Account Service API'
  description: 'API USED TO RETURN INFORMATION TO THE ODDS APPLICATION'
  version: '1.0.0'
servers:
  - url: 'http://localhost:8080/api/v1/account'

paths:
  /all:
    get:
      description: 'Returns a list of all available accounts'
      operationId: 'getAllAccounts'
      responses:
        '200':
          description: 'Successful response from API'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsResponse'
        '400':
          description: 'Request is invalid: Either required data is missing or data is not consistent'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'Authorization Failure. Customer token does not match'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'Not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 'Service Unavailable'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /{id}:
    get:
      description: 'Return an account matched with the id'
      operationId: 'getAccountById'
      parameters:
        - name: 'id'
          in: 'path'
          required: true
          schema:
            type: number
          description: Determines which account is to be returned
      responses:
        '200':
          description: 'Successful response from API'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountItem'
        '400':
          description: 'Request is invalid: Either required data is missing or data is not consistent'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'Authorization Failure. Customer token does not match'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'Not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: 'Service Unavailable'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    AccountItem:
      type: 'object'
      description: 'Basic account information returned'
      properties:
        firstName:
          description: 'First name associated with the account'
          type: 'string'
          maxLength: 20
          example: 'Jane'
        lastName:
          description: 'Last name associated with the account'
          type: 'string'
          maxLength: 20
          example: 'Doe'
        email:
          description: 'Email associated with the account'
          type: 'string'
          maxLength: 50
          example: 'email@example.com'
        phoneNumber:
          description: 'Phone number associated with account'
          type: 'string'
          maxLength: 12
          example: '000-000-0000'
        bankAccountNumber:
          description: 'Bank account number associated with the account'
          type: 'string'
          maxLength: 20
          example: '000000000'
        id:
          description: 'Account number associated with the account'
          type: 'string'
          maxLength: 10
          example: 2
    AccountsResponse:
      type: 'object'
      description: 'List of available accounts returned'
      properties:
        accountItem:
          type: 'array'
          items:
            $ref: '#/components/schemas/AccountItem'
    Error:
      type: 'object'
      description: 'Object that contains detailed error information'
      properties:
        errorCode:
          type: 'string'
          maxLength: 3
          description: 'Failure/issue code'
          example: '200'
        errorDescription:
          type: 'string'
          maxLength: 250
          description: 'Freeform explanatory text for the error'
          example: 'Authorization failure. Either the Authorization or the Authorization value did not match the expected value'
        errorCategoryName:
          type: 'string'
          maxLength: 15
          description: 'Name/decode for the http response status code class. Ex: 1xx(informational)'
          example: 'APPLICATION'